Confirmacion para opcion eliminar

+++ WebContent/app/lista_categoria.jsp
@@ -40,7 +40,7 @@
 			                <display:column  title="Acciones" sortable="false">
 			                  <div class="btn-group btn-group-sm" role="group" aria-label="...">
 			                    <a href="ServletCategoria?tipo=buscar&cod=${lista.codigo}" class="btn btn-default" title="Editar"><i class="fa fa-pencil" aria-hidden="true"></i></a>
-			                    <a href="ServletCategoria?tipo=eliminar&cod=${lista.codigo}" class="btn btn-default" title="Eliminar"><i class="fa fa-trash-o" aria-hidden="true"></i></a>
+			                    <a onclick="javascript:if(!confirm('Por favor, confirme su elección')) return false;" href="ServletCategoria?tipo=eliminar&cod=${lista.codigo}" class="btn btn-default" title="Eliminar"><i class="fa fa-trash-o" aria-hidden="true"></i></a>
 			                  </div>
 				            </display:column>
 				        </display:table>
